# backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  pointmarket_backend_internal_dtos.APIError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  pointmarket_backend_internal_dtos.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  pointmarket_backend_internal_dtos.AssignmentDTO:
    properties:
      course_id:
        type: integer
      course_title:
        type: string
      created_at:
        type: string
      description:
        description: Use pointer for nullable
        type: string
      due_date:
        description: Use pointer for nullable
        type: string
      id:
        type: integer
      reward_points:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  pointmarket_backend_internal_dtos.CreateAssignmentRequestDTO:
    properties:
      course_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      reward_points:
        type: integer
      status:
        description: e.g., "draft", "published"
        type: string
      title:
        type: string
    required:
    - course_id
    - reward_points
    - title
    type: object
...
  pointmarket_backend_internal_dtos.CreateStudentAssignmentRequestDTO:
    properties:
      assignment_id:
        type: integer
      status:
        description: e.g., "not_started", "in_progress"
        type: string
      student_id:
        type: integer
      submission:
        type: string
    required:
    - assignment_id
    - student_id
    type: object
...
  pointmarket_backend_internal_dtos.ListAssignmentsResponseDTO:
    properties:
      assignments:
        items:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.AssignmentDTO'
        type: array
      total:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.ListStudentAssignmentsResponseDTO:
    properties:
      student_assignments:
        items:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentAssignmentDTO'
        type: array
      total:
        type: integer
    type: object
...
  pointmarket_backend_internal_dtos.NullData:
    type: object
...
  pointmarket_backend_internal_dtos.UpdateAssignmentRequestDTO:
    properties:
      course_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      reward_points:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: PointMarket backend API documentation.
  title: PointMarket API
  version: "1.0"
paths:
...
  /assignments:
    get:
      description: Retrieves a list of all assignments, with optional filtering by
        course ID.
      parameters:
      - description: Filter by Course ID
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.ListAssignmentsResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get all assignments
      tags:
      - assignments
    post:
      consumes:
      - application/json
      description: Creates a new assignment with the provided details.
      parameters:
      - description: Assignment details
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateAssignmentRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.AssignmentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Create a new assignment
      tags:
      - assignments
  /assignments/{id}:
    delete:
      description: Deletes an assignment by its ID.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Delete an assignment
      tags:
      - assignments
    get:
      description: Retrieves a single assignment by its unique identifier.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.AssignmentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get assignment by ID
      tags:
      - assignments
    put:
      consumes:
      - application/json
      description: Updates an existing assignment identified by its ID.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated assignment details
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpdateAssignmentRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.AssignmentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Update an assignment
      tags:
      - assignments
  /assignments/{id}/start:
    post:
      consumes:
      - application/json
      description: Records that a student has started a specific assignment.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional payload; assignment_id and student_id are derived
        in: body
        name: studentAssignment
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.CreateStudentAssignmentRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentAssignmentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Record student starting an assignment
      tags:
      - assignments
  /assignments/{id}/submissions:
    get:
      description: Retrieves all student assignment records for a specific assignment.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.ListStudentAssignmentsResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Get assignment submissions
      tags:
      - assignments
  /assignments/{id}/submissions/{student_assignment_id}:
    delete:
      description: Deletes a student's assignment record by its ID.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student Assignment ID
        in: path
        name: student_assignment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.NullData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Delete an assignment submission
      tags:
      - assignments
    put:
      consumes:
      - application/json
      description: Students submit via `/assignments/{id}/submit`; teachers/admin
        grade via `/assignments/{id}/submissions/{student_assignment_id}`.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student Assignment ID (grading route only)
        in: path
        name: student_assignment_id
        type: integer
      - description: Student assignment update
        in: body
        name: studentAssignment
        required: true
        schema:
          $ref: '#/definitions/pointmarket_backend_internal_dtos.UpdateStudentAssignmentRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pointmarket_backend_internal_dtos.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/pointmarket_backend_internal_dtos.StudentAssignmentDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pointmarket_backend_internal_dtos.APIError'
      security:
      - BearerAuth: []
      summary: Submit an assignment or grade a submission
      tags:
      - assignments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"